evaluator/pyproject.toml
[project]
name = "evaluator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.5",
    "deepeval>=3.7.2",
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-ollama>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "milvus-lite>=2.5.1",
    "pymilvus>=2.6.3",
    "pymilvus-model>=0.3.2",
    "pypdf>=6.2.0",
    "python-dotenv>=1.2.1",
    "ragas>=0.3.9",
    "rapidfuzz>=3.14.3",
    "scipy>=1.16.3",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
]
